requirements.txt
numpy~=1.20.3
scikit-learn~=0.24.2
matplotlib~=3.4.2
pandas~=1.2.4
scikit-multilearn~=0.2.0
tqdm~=4.61.0
scipy~=1.6.3
seaborn~=0.11.1
PyYAML~=5.4.1